+++ src/relations/postgresql_provider.py
@@ -6,7 +6,7 @@
 
 import logging
 
-from charms.data_platform_libs.v0.database_provides import (
+from charms.data_platform_libs.v0.data_interfaces import (
     DatabaseProvides,
     DatabaseRequestedEvent,
 )
