+++ tests/integration/ha_tests/application-charm/src/charm.py
@@ -15,7 +15,7 @@ import subprocess
 from typing import Dict, Optional
 
 import psycopg2
-from charms.data_platform_libs.v0.database_requires import DatabaseRequires
+from charms.data_platform_libs.v0.data_interfaces import DatabaseRequires
 from ops.charm import ActionEvent, CharmBase
 from ops.main import main
 from ops.model import ActiveStatus, Relation
