+++ tests/integration/new_relations/application-charm/src/charm.py
@@ -10,7 +10,7 @@ of the libraries in this repository.
 
 import logging
 
-from charms.data_platform_libs.v0.database_requires import (
+from charms.data_platform_libs.v0.data_interfaces import (
     DatabaseCreatedEvent,
     DatabaseEndpointsChangedEvent,
     DatabaseRequires,
